Start work on parallel tests
